rename the package name to 'queryfixer'
add documents for classes
