+++ tools/automatic_query_fixer/src/main/java/com/google/cloud/bigquery/utils/queryfixer/BigQueryParserFactory.java
@@ -1,4 +1,4 @@
-package com.google.cloud.bigquery.utils.autoqueryfixer;
+package com.google.cloud.bigquery.utils.queryfixer;
 
 import org.apache.calcite.avatica.util.Casing;
 import org.apache.calcite.avatica.util.Quoting;
@@ -11,7 +11,9 @@ import org.apache.calcite.util.SourceStringReader;
 
 import java.io.Reader;
 
-
+/**
+ * A factory to generate parsers. The fault generated parser is Babel Parser with BigQuery dialect.
+ * */
 public class BigQueryParserFactory {
 
   private final SqlParserImplFactory factory;
@@ -21,20 +23,35 @@ public class BigQueryParserFactory {
   private final SqlConformance conformance = SqlConformanceEnum.DEFAULT;
   private final SqlParser.Config parserConfig;
 
+  /**
+   * Default initialization with the Babel Parser factory.
+   * */
   public BigQueryParserFactory() {
     factory = SqlBabelParserImpl.FACTORY;
     this.parserConfig = buildConfig();
   }
 
+  /**
+   * Initialization with customized factory
+   * @param factory customized factory */
   public BigQueryParserFactory(SqlParserImplFactory factory) {
     this.factory = factory;
     this.parserConfig = buildConfig();
   }
 
+  /**
+   * Get the parser parsing the input query.
+   * @param sql query to parse
+   * @return a parser loaded with the query
+   */
   public SqlParser getParser(String sql) {
     return getParser(new SourceStringReader(sql));
   }
 
+  /**
+   * Get the config of the parser factory
+   * @return parser config
+   */
   public SqlParser.Config getParserConfig() {
     return parserConfig;
   }
