+++ tools/automatic_query_fixer/src/main/java/com/google/cloud/bigquery/utils/queryfixer/entity/ChoiceRequest.java
@@ -1,4 +1,4 @@
-package com.google.cloud.bigquery.utils.autoqueryfixer.entity;
+package com.google.cloud.bigquery.utils.queryfixer.entity;
 
 import lombok.Data;
 
@@ -6,20 +6,18 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * A request to ask users to choose/confirm an option to fix the query.
- * The request contains error position and a list of fix options.
+ * A request to ask users to choose/confirm an option to fix the query. The request contains error
+ * position and a list of fix options.
+ *
  * @see package com.google.cloud.bigquery.utils.auto_query_fixer.entity.FixOption
  */
-
 @Data
 public class ChoiceRequest {
 
   Object errPos;
   List<FixOption> fixOptions;
 
-  /**
-   * create a choiceRequest and initialize its fixOptions List.
-   * */
+  /** create a choiceRequest and initialize its fixOptions List. */
   public ChoiceRequest() {
     fixOptions = new ArrayList<>();
   }
