+++ tools/automatic_query_fixer/src/main/java/com/google/cloud/bigquery/utils/queryfixer/entity/FixResult.java
@@ -1,4 +1,4 @@
-package com.google.cloud.bigquery.utils.autoqueryfixer.entity;
+package com.google.cloud.bigquery.utils.queryfixer.entity;
 
 
 import lombok.Data;
