+++ tools/automatic_query_fixer/src/main/java/com/google/cloud/bigquery/utils/queryfixer/entity/TokenImpl.java
@@ -1,4 +1,4 @@
-package com.google.cloud.bigquery.utils.autoqueryfixer.entity;
+package com.google.cloud.bigquery.utils.queryfixer.entity;
 
 import org.apache.calcite.sql.parser.babel.Token;
